pandas frontend Series class logic and abs methods refactored to the generic class NDFrame to allow method sharing between Series and DataFrame as well as the common logic. this is loosely based on the super class in pandas code base https://github.com/pandas-dev/pandas/blob/0f437949513225922d851e9581723d82120684a6/pandas/core/generic.py#L226 for both the Series and DataFrame
