+++ ivy/functional/frontends/pandas/__init__.py
@@ -1,2 +1,4 @@
 from . import series
 from .series import *
+from . import generic
+from .generic import *
